Replace Message#send with Message#reply
Maintains the same behavior as Message#send, but is a better name for what it does once you exclude the preconceived atrocity from djs
