+++ src/commands/Admin/unload.ts
@@ -16,10 +16,10 @@ export default class extends Command {
 
 	public async run(message: Message, [piece]: [Piece]): Promise<Message[]> {
 		if ((piece.type === 'event' && piece.name === 'message') || (piece.type === 'monitor' && piece.name === 'commandHandler')) {
-			return message.sendLocale('COMMAND_UNLOAD_WARN');
+			return message.replyLocale('COMMAND_UNLOAD_WARN');
 		}
 		piece.unload();
-		return message.sendLocale('COMMAND_UNLOAD', [piece.type, piece.name]);
+		return message.replyLocale('COMMAND_UNLOAD', [piece.type, piece.name]);
 	}
 
 }
