+++ src/events/argumentError.ts
@@ -5,7 +5,7 @@ import type { Argument } from 'klasa';
 export default class extends Event {
 
 	public async run(message: Message, _argument: Argument, _params: readonly unknown[], error: string): Promise<void> {
-		await message.send(mb => mb.setContent(error));
+		await message.reply(mb => mb.setContent(error));
 	}
 
 }
