+++ src/monitors/commandHandler.ts
@@ -17,7 +17,7 @@ export default class CommandHandler extends Monitor {
 		if (!message.channel.postable) return undefined;
 		if (!message.commandText && message.prefix === this.client.mentionPrefix) {
 			const prefix = message.guildSettings.get('prefix') as string | string[];
-			return message.sendLocale('PREFIX_REMINDER', [prefix.length ? prefix : undefined]);
+			return message.replyLocale('PREFIX_REMINDER', [prefix.length ? prefix : undefined]);
 		}
 		if (!message.commandText) return undefined;
 		if (!message.command) {
